Added new decideLoot function
Added a "decideLoot()" function to decide whether or not to pick up the new item. If the item is too weak, don't pick it up. If the item is stronger than yours, pick it up. If the item type (ie. armor or damage item) doesn't exist in your inventory, pick it up. (It just returns true or false).